document-qa-chatbot/requirements.txt.txt
streamlit
pypdf
sentence-transformers
langchain-text-splitters
langchain-google-genai
chromadb
python-dotenv